fix(documents/enriched): exclude EnrichedFigure from 'enrichers'
The EnrichedFigure class is a base class for all enriched Figure
classes, which are obtained by the get_figure_enriching_classes()
function. However, this function overlooked the fact that "all classes
are subclasses of themselves," leading to the unintentional inclusion of
the base class.

This change fixes the issue by explicitly checking whether the given
class is the base class (EnrichedFigure).

Signed-off-by: Morten B. Rasmussen <m.rasmussen@samsung.com>