feat(documents/model): add an initial Model document
The intent of the model document is to provide a non-figure bound
representation of specification concepts, in a form convenient for an
implementer.

Thus, data from multiple figures combined into representing a "Command",
this grabs data from opcode-figures, support-requirements-figures,
command-sqe/cqe-dword figures etc. with such a representation, the task
of maintaining drivers and tooling will be dramatically simplified.

The initial consumer of this is the xNVMe project.

Signed-off-by: Simon A. F. Lund <os@safl.dk>