refactor: move enrichment into converter
The implementation of "enrichment" and partly "conversion" was scattered
accross the EnrichmentDocument Converter (FromFigureDocument) and the
Grid class-method. This made it slightly cumbersome for data-access
in the Grid class-method as it was supposed to only have access to the
table, but it needed the entire Figure.

Thus, seems like a better code organization to keep all the conversion
logic, inside the converter. This is now done using a handful of
helpers, marked static.

Signed-off-by: Simon A. F. Lund <os@safl.dk>