fix(enriched): change varying row length to blocking error
We see examples of tables being incorrectly parsed, because its are
given as:
Header: [Bytes, Description, Description]
Rows: [Bytes, Bytes, Description, Description]

For example figure 228 in the base specifications. Here the two Bytes
columns in the table data rows are parsed as both the Bytes and
Description and the Description columns are completely disregarded.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>